requirements.txt
pandas==2.2.1
numpy==1.26.4
python==3.11
nltk==3.8.1

PyMuPDF==1.24.2
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
spacy==3.7.4
transformers==4.38.2
FAISS-cpu==1.8.0

